requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
requests==2.31.0
Werkzeug==3.0.1
python-dotenv==1.0.0
openmeteo-requests
requests-cache
retry-requests
numpy
